fix: refactor friends APIs
