+++ src/api/friend/friend.module.ts
@@ -4,10 +4,16 @@ import { FriendController } from './controller/friend.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { FriendEntity } from './model/friend.entity';
 import { FriendGroupEntity } from '../friend-group/model/friend-group.entity';
+import { FriendGroupHandlerService } from './service/repository-handler/friend-group-handler.service';
+import { FriendListHandlerService } from './service/repository-handler/friend-list-handler.service';
 
 @Module({
   controllers: [FriendController],
-  providers: [FriendService],
+  providers: [
+    FriendService,
+    FriendGroupHandlerService,
+    FriendListHandlerService,
+  ],
   imports: [TypeOrmModule.forFeature([FriendEntity, FriendGroupEntity])],
 })
 export class FriendModule {}
