+++ src/api/friend/service/friend.service.ts
@@ -1,23 +1,23 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { Injectable } from '@nestjs/common';
 import { CreatedFriendReponse, FriendResponse } from '../model/friend.response';
-import { FriendEntity } from '../model/friend.entity';
-import { FriendGroupEntity } from '../../friend-group/model/friend-group.entity';
 import { FriendDto } from '../model/friend.dto';
+import { FriendGroupHandlerService } from './repository-handler/friend-group-handler.service';
+import { FriendListHandlerService } from './repository-handler/friend-list-handler.service';
 
 @Injectable()
 export class FriendService {
   constructor(
-    @InjectRepository(FriendEntity)
-    private friendListEntityRepository: Repository<FriendEntity>,
-    @InjectRepository(FriendGroupEntity)
-    private friendGroupEntityRepository: Repository<FriendGroupEntity>,
+    private readonly friendGroupHandlerService: FriendGroupHandlerService,
+    private readonly friendListHandlerService: FriendListHandlerService,
   ) {}
 
-  async getFriends(groupId: number): Promise<FriendResponse> {
-    const friendInfo = await this.getFriendInfo(groupId);
-    const groupName = await this.getGroupName(groupId);
+  async getFriendsWith(groupId: number): Promise<FriendResponse> {
+    const friendInfo = await this.friendListHandlerService.getFriendsWith(
+      groupId,
+    );
+    const groupName = await this.friendGroupHandlerService.getGroupName(
+      groupId,
+    );
     return {
       groupId,
       groupName,
@@ -26,13 +26,7 @@ export class FriendService {
   }
 
   async setFriend(dto: FriendDto): Promise<CreatedFriendReponse> {
-    const friendEntityForSave = this.getFriendEntityForSave(dto);
-    const { id } = await this.friendListEntityRepository.save(
-      friendEntityForSave,
-    );
-    if (!id) {
-      throw new InternalServerErrorException('저장에 실패했습니다.');
-    }
+    const id = await this.friendListHandlerService.saveFriend(dto);
     return { friendId: id };
   }
 
@@ -40,42 +34,7 @@ export class FriendService {
     friendId: number,
     dto: FriendDto,
   ): Promise<CreatedFriendReponse> {
-    try {
-      await this.friendListEntityRepository.update(friendId, dto);
-    } catch (e) {
-      throw new InternalServerErrorException('친구정보 수정에 실패했습니다.');
-    }
+    await this.friendListHandlerService.updateFriend(friendId, dto);
     return { friendId };
-  }
-
-  private getFriendEntityForSave(dto: FriendDto) {
-    return {
-      isMember: !!dto.kakaoId,
-      ...dto,
-    } as FriendEntity;
-  }
-
-  private async getGroupName(groupId: number) {
-    const group = await this.friendGroupEntityRepository.findOneBy({
-      id: groupId,
-    });
-
-    if (!group?.name) {
-      throw new InternalServerErrorException('그룹을 찾을 수 없습니다.');
-    }
-    return group.name;
-  }
-
-  private async getFriendInfo(groupId: number) {
-    const friends = await this.friendListEntityRepository.findBy({
-      groupId,
-    });
-    return friends.map((friend) => {
-      return {
-        id: friend.id,
-        name: friend.name,
-        thumbnailImageUrl: friend.thumbnailImageUrl,
-      };
-    });
   }
 }
